refactor: 将 agentModes 从 constants/ 迁移至 utils/ 以符合模块分层规范
