+++ src/renderer/utils/agentModes.ts
